refactor(media): extract the streaming layer into CwmmediaStreamer (#1778)
CwmpodcastTrackHelper carried a complete HTTP file streamer — Range,
206/416, Accept-Ranges, Last-Modified/304, HEAD, and a proxy path for
external hosts with an SSRF guard. None of it was podcast-specific; it
only lived there because the podcast download-tracking redirect (#1424)
was the first thing to need it.

The front-end download route needs the same behaviour (#1774), and
Cwmdownload currently sends Content-Length + readfile() with no Range
support at all, so gated audio and video could not be seeked. Writing
that a second time would mean two implementations of the same protocol
handling, one of which has already been through an SSRF patch.

So the ten generic methods move to CwmmediaStreamer verbatim —
serveMedia (as serve), resolveLocalPath, streamLocalFile,
streamRemoteFile, isLocalHost, resolveSafeRemoteIp, resolveRange,
flushBuffers, terminate, fail. The podcast-specific half stays put: bot
detection, client hashing, GUID resolution, media lookup and hit
recording.

Behaviour-neutral. The only new logic line is the delegate call;
everything else is relocation, which the diff shows. serveMedia() is
kept as a wrapper because the podcast redirect is its own entry point
and reads better named for what it does there.

Tests follow the code, as they did when this logic moved out of the
controller in #1446: the reflections on isLocalHost, streamRemoteFile
and resolveRange now target CwmmediaStreamer, including the source-level
guard that every streaming branch terminates rather than returning. The
podcast-specific tests still point at CwmpodcastTrackHelper. 66 tests
pass unchanged.

⚠️ The class docblock states what the old caller never had to: this
answers how to send a file, never whether to. serveMedia() performs no
access check by design — Apple's crawler has no session — so callers
gating content must check first, per Cwmdownload::isAccessible() (#1777).

The SSRF guard from 10.5.5 (#1426) moved with the remote path rather
than being left behind; asserted present in the new file.

Refs #1774
